refactor: Directory structure of domain
