+++ test/unit/domain/consents/generateChallenge.test.ts
@@ -28,19 +28,19 @@
  --------------
  ******/
 import { Request } from '@hapi/hapi'
-import { consentDB } from '../../../../../src/lib/db'
-import { Consent } from '../../../../../src/model/consent'
-import { thirdPartyRequest } from '../../../../../src/lib/requests'
+import { consentDB } from '../../../../src/lib/db'
+import { Consent } from '../../../../src/model/consent'
+import { thirdPartyRequest } from '../../../../src/lib/requests'
 // eslint-disable-next-line max-len
 import {
   updateConsentCredential,
   isConsentRequestInitiatedByValidSource,
   putConsentId
-} from '../../../../../src/domain/consents/{ID}/generateChallenge'
+} from '../../../../src/domain/consents/generateChallenge'
 import { Enum } from '@mojaloop/central-services-shared'
 import Logger from '@mojaloop/central-services-logger'
 import { GenericRequestResponse } from '@mojaloop/sdk-standard-components'
-import { ExternalScope } from '../../../../../src/lib/scopes'
+import { ExternalScope } from '../../../../src/lib/scopes'
 
 // Declaring Mock Functions
 const mockPutConsents = jest.spyOn(thirdPartyRequest, 'putConsents')
