+++ test/unit/server/handlers/consents/{ID}/generateChallenge.test.ts
@@ -33,7 +33,7 @@ import { consentDB, scopeDB } from '../../../../../../src/lib/db'
 import { Consent } from '../../../../../../src/model/consent'
 import * as Handler from '../../../../../../src/server/handlers/consents/{ID}/generateChallenge'
 import * as Challenge from '../../../../../../src/lib/challenge'
-import * as Domain from '../../../../../../src/domain/consents/{ID}/generateChallenge'
+import * as Domain from '../../../../../../src/domain/consents/generateChallenge'
 import * as ScopeFunctions from '../../../../../../src/lib/scopes'
 import Logger from '@mojaloop/central-services-logger'
 import { Scope } from '../../../../../../src/model/scope'
